# platform/api-server/src/domains/early-warnings/handlers/early-warning.handlers.ts
/**
 * Handlers for EarlyWarning
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: early-warnings
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { EarlyWarningsDomainModule } from "../dependencies/early-warnings-ddd.dependencies.js";
import type { GetEarlyWarningInput, ListEarlyWarningsInput } from "@lendora/services/early-warnings";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/early-warnings
 * Operation: listEarlyWarnings
 */
export async function listEarlyWarnings(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], status?: string | string[], severity?: string | string[], vintageWindow?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: EarlyWarningsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListEarlyWarningsInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.earlyWarnings.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for GET /v1/early-warnings/{alertId}
 * Operation: getEarlyWarning
 */
export async function getEarlyWarning(
  request: FastifyRequest<{ Params: { alertId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: EarlyWarningsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetEarlyWarningInput = {
    alertId: request.params.alertId,
  };
  const result = await deps.useCases.earlyWarnings.get.execute(input);
  return reply.code(200).send(result);
}