# platform/api-server/src/domains/policies/handlers/policy.handlers.ts
/**
 * Handlers for Policy
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: policies
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { PoliciesDomainModule } from "../dependencies/policies-ddd.dependencies.js";
import type { CreatePolicyPackInput, GetPolicyPackInput, ListPolicyPacksInput, UpdatePolicyPackInput } from "@lendora/services/policies";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/policies
 * Operation: listPolicyPacks
 */
export async function listPolicyPacks(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], productLine?: string | string[], status?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: PoliciesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListPolicyPacksInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.policies.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/policies
 * Operation: createPolicyPack
 */
export async function createPolicyPack(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: PoliciesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: CreatePolicyPackInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.policies.create.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/policies/{policyPackId}
 * Operation: getPolicyPack
 */
export async function getPolicyPack(
  request: FastifyRequest<{ Params: { policyPackId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: PoliciesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetPolicyPackInput = {
    policyPackId: request.params.policyPackId,
  };
  const result = await deps.useCases.policies.get.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for PATCH /v1/policies/{policyPackId}
 * Operation: updatePolicyPack
 */
export async function updatePolicyPack(
  request: FastifyRequest<{ Params: { policyPackId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: PoliciesDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: UpdatePolicyPackInput = {
    policyPackId: request.params.policyPackId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.policies.update.execute(input);
  return reply.code(200).send(result);
}