# platform/api-server/src/domains/referrals/handlers/refer-task.handlers.ts
/**
 * Handlers for ReferTask
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: referrals
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { ReferralsDomainModule } from "../dependencies/referrals-ddd.dependencies.js";
import type { GetReferTaskInput, ListReferTasksInput } from "@lendora/services/referrals";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/refer-tasks
 * Operation: listReferTasks
 */
export async function listReferTasks(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], status?: string | string[], productLine?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: ReferralsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListReferTasksInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.referTasks.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for GET /v1/refer-tasks/{referTaskId}
 * Operation: getReferTask
 */
export async function getReferTask(
  request: FastifyRequest<{ Params: { referTaskId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: ReferralsDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetReferTaskInput = {
    referTaskId: request.params.referTaskId,
  };
  const result = await deps.useCases.referTasks.get.execute(input);
  return reply.code(200).send(result);
}